// shop-admin-frontend/src/@core/libs/form-builder/interfaces/datatype.interface.ts
export enum ControlType {
	Boolean = 1,
	Input = 2,
	Select = 3,
	Image = 4,
	DateTime = 5,
	Radio = 6,
	NumberInt = 7,
	SelectMultiple = 8,
	SelectMultipleTree = 9,
	CheckBox = 10,
	multiRows = 11,
	LongText = 12,
	QuillEditor = 13,
	UploadFile = 14,
	Email = 20,
	Formula = 21,
	Table = 22,
	Currency = 23,
	Data = 24,
	SubTask = 25,
	People = 26,
	Progress = 27,
	Rating = 28
}

export enum DateTimeFormat {
	MONTH_YEAR = 1,
	YEAR = 2,
	DAY_MONTH_YEAR_TIME = 4,
	HOUR_MINUTE = 5,
	DAY_MONTH_YEAR = 6
}